Update cursor after null correctly
For a number of types of field value contained within an object, a
null value would result in errors of the form:

Invalid JSON, wrong char 'l' found at position xx

This appears to be due to the cursor not consistently being updated
after the null value is detected.

This patch resolves the issue by consistently updating the cursor in
each case.