Fix interface decoding when string has escape sequence
Decoding of a string using the interface API fails when a string has an escape
sequence. Go's native JSON decoding function is called in such a case, and it
expects to get the original string. Provide the raw string to prevent the
failures that are currently caused.

Adding a test for this specific scenario.

Signed-off-by: Nadav Amit <namit@vmware.com>